moved pipeline tests
